Проведена модуляция
