+++ js/advertisements-data.js
@@ -1,0 +1,37 @@
+const TYPES = [
+  'palace',
+  'flat',
+  'house',
+  'bungalow',
+];
+
+const TIME = [
+  '12:00',
+  '13:00',
+  '14:00',
+];
+
+const FEATURES = [
+  'wifi',
+  'dishwasher',
+  'parking',
+  'washer',
+  'elevator',
+  'conditioner',
+];
+
+const PHOTOS = [
+  'http://o0.github.io/assets/images/tokyo/hotel1.jpg',
+  'http://o0.github.io/assets/images/tokyo/hotel2.jpg',
+  'http://o0.github.io/assets/images/tokyo/hotel3.jpg',
+];
+
+const DESCRIPTIONS = [
+  'Светло,уютно и идеально для отдыха и командировок!',
+  'Проведите незабываемые выходные с комфортом!',
+  'Комфортное жильё по доступной цене!',
+];
+
+const ADVERTISEMENTS_COUNT = 10;
+
+export {TYPES, TIME, FEATURES, PHOTOS, DESCRIPTIONS, ADVERTISEMENTS_COUNT};
